[r1 #034] Accept fractional HUD episode timeouts
(cherry picked from commit 71d89b7c9c7e11c9112a0c610d0b96d0b3b50f75)

+++ examples/experimental/hud/rollout.py
@@ -61,7 +61,7 @@ class HudConfig(BaseConfig):
     hud_shot_width: A[int, Arg()] = 512
     hud_max_tokens_per_turn: A[int, Arg()] = 512
     hud_max_sandboxes: A[int, Arg()] = 32
-    hud_episode_timeout_s: A[int, Arg()] = 1800
+    hud_episode_timeout_s: A[float, Arg()] = 1800.0
     hud_daytona_key_file: A[str | None, Arg()] = None
     hud_system_prompt: A[str | None, Arg()] = None
